Generalize the GitHubInfoProvider
