--- src/web/src/hooks/use-githubInfo.ts
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-import { GithubInfoContext, GithubInfoContextInterface } from '../components/GithubInfoProvider';
-
-const useGithubInfo = (): GithubInfoContextInterface => useContext(GithubInfoContext);
-
-export default useGithubInfo;
